# Speedup ExactSampler